Added BiblioWoods Addons
BiblioWoods Addons

These add-ons for BiblioCraft will allow the player to craft 8 of the
wooden BiblioCraft blocks in the different woods provided by said mod.
For any BiblioCraft block that currently can be specified by the 4
vanillas woods can be made the same way with the modded woods and you
will receive a block of that color wood.